Create login blueprint
